+++ admin/frontend/components/sidebar.js
@@ -10,8 +10,9 @@ import Organizations from '../pages/organizations/list'
 import Roles from '../pages/roles/list'
 import Groups from '../pages/groups/list'
 
-import RequestLogs from '../pages/request-logs/list'
+import RequestLogs from '../pages/developer-tools/request-logs'
 import AppConfig from '../pages/developer-tools/app-config'
+import FormBuilder from '../pages/developer-tools/form-builder'
 // #Import
 
 class Sidebar extends Component {
@@ -73,7 +74,8 @@ class Sidebar extends Component {
         open: false,
         dropdown: [
           RequestLogs.asSidebarItem(),
-          AppConfig.asSidebarItem()
+          AppConfig.asSidebarItem(),
+          FormBuilder.asSidebarItem()
         ]
       }, {
         title: 'Restore data',